+++ tests/acceptance/lute_test_client.py
@@ -15,6 +15,7 @@ to get nicer assertion details.
 """
 
 import time
+import json
 import requests
 from selenium.webdriver.common.keys import Keys
 from selenium.webdriver.common.action_chains import ActionChains
@@ -130,13 +131,50 @@ class LuteTestClient:  # pylint: disable=too-many-public-methods
         "Get book table content."
         css = "#booktable tbody tr"
 
+        # Skip the last two columns:
+        # - "last opened" date is a hassle to check
+        # - "actions" is just "..."
         def _to_string(row):
             tds = row.find_by_css("td")
             rowtext = [td.text.strip() for td in tds]
-            return "; ".join(rowtext).strip()
+            ret = "; ".join(rowtext[:-2]).strip()
+            # print(ret, flush=True)
+            return ret
 
         rows = list(self.browser.find_by_css(css))
         return "\n".join([_to_string(row) for row in rows])
+
+    def get_book_page_start_dates(self):
+        "get content from sql check"
+        sql = """select bktitle, txorder
+        from books
+        inner join texts on txbkid = bkid
+        where txstartdate is not null
+        order by bktitle, txorder"""
+        response = requests.get(f"{self.home}/dev_api/sqlresult/{sql}", timeout=1)
+        ret = "\n".join(json.loads(response.text))
+        if ret == "":
+            ret = "-"
+        return ret
+
+    def get_book_page_read_dates(self):
+        "get content from sql check"
+        sql = """select bktitle, txorder
+        from books
+        inner join texts on txbkid = bkid
+        where txreaddate is not null
+        order by bktitle, txorder"""
+        response = requests.get(f"{self.home}/dev_api/sqlresult/{sql}", timeout=1)
+        ret = "\n".join(json.loads(response.text))
+        if ret == "":
+            ret = "-"
+        return ret
+
+    def set_txstartdate_to_null(self):
+        "hack back end to keep test data sane."
+        sql = "update texts set txstartdate = null"
+        response = requests.get(f"{self.home}/dev_api/execsql/{sql}", timeout=1)
+        return response.text
 
     ################################
     # Terms
@@ -242,11 +280,16 @@ class LuteTestClient:  # pylint: disable=too-many-public-methods
 
     def make_term(self, lang, updates):
         "Create a new term."
-        self.visit("/")
-        self.browser.find_by_css("#menu_terms").mouse_over()
-        self.browser.find_by_id("term_index").first.click()
-        self.browser.find_by_css("#term_actions").mouse_over()
-        self.click_link("Create new")
+        # Sometimes this failed during the unsupported_parser.feature, not sure why.
+        # Likely something silly, don't care, so will bypass the screen controls.
+        # I am sure this will bite me later.
+        # TODO fix_nav: figure out why occasionally got ElementNotInteractableException
+        # self.visit("/")
+        # self.browser.find_by_css("#menu_terms").mouse_over()
+        # self.browser.find_by_id("term_index").first.click()
+        # self.browser.find_by_css("#term_actions").mouse_over()
+        # self.click_link("Create new")
+        self.visit("/term/new")
         assert "New Term" in self.browser.html
 
         updates["language_id"] = self.language_ids[lang]
@@ -369,6 +412,33 @@ class LuteTestClient:  # pylint: disable=too-many-public-methods
         actions.release()
         actions.perform()
 
+    def _refresh_browser(self):
+        """
+        Term actions (edits, hotkeys) cause updated content to be ajaxed in.
+        For the splinter browser to be aware of it, the browser has to be
+        reloaded, but calling a self.browser.reload() has other side effects
+        (sets the page start date, etc).
+
+        The below weird js hack causes the browser to be updated,
+        and then the js events have to be reattached too.
+        """
+        # self.browser.reload()
+        # ??? ChatGPT suggested:
+        time.sleep(0.5)  # Hack for ci.
+        self.browser.execute_script(
+            """
+            // Trigger re-render of the entire body
+            var body = document.querySelector('body');
+            var content = body.innerHTML;
+            body.innerHTML = '';
+            body.innerHTML = content;
+
+            // Re-attach text interactions.
+            window.prepareTextInteractions();
+            """
+        )
+        time.sleep(0.5)  # Hack for ci.
+
     def fill_reading_bulk_edit_form(self, updates=None):
         """
         Click a word in the reading frame, fill in the term form iframe.
@@ -388,11 +458,20 @@ class LuteTestClient:  # pylint: disable=too-many-public-methods
             if "updated" in iframe.html:
                 should_refresh = True
 
-        # Have to refresh the content to query the dom ...
-        # Unfortunately, I can't see how to refresh without reloading
+        # Have to refresh the content to query the dom.
         if should_refresh:
-            self.browser.reload()
-            time.sleep(0.2)  # Hack, test failing.
+            self._refresh_browser()
+
+    def hack_set_hotkey(self, hotkey, value):
+        "Hack set hotkey directly through dev api.  Trashy."
+        sql = f"""update settings
+        set StValue='{value}' where StKey='{hotkey}'"""
+        requests.get(f"{self.home}/dev_api/execsql/{sql}", timeout=1)
+        # NOTE! Hacking is dumb, it bypassing the global state which is rendered in JS.
+        # Have to visit and save settings to re-set the JS values that will be rendered.
+        # Big time waste finding this out.
+        self.visit("settings/shortcuts")
+        self.browser.find_by_css("#btnSubmit").first.click()
 
     def press_hotkey(self, hotkey):
         "Send a hotkey."
@@ -409,27 +488,29 @@ class LuteTestClient:  # pylint: disable=too-many-public-methods
             "i": "KeyI",
             "m": "KeyM",
             "w": "KeyW",
+            # Manually added.
+            "8": "Digit8",
+            "9": "Digit9",
         }
-        if hotkey not in key_to_code_map:
+        if hotkey.lower() not in key_to_code_map:
             raise RuntimeError(f"Missing {hotkey} in acceptance test map")
         event_parts = [
             "type: 'keydown'",
-            f"code: '{key_to_code_map[hotkey]}'",
+            f"code: '{key_to_code_map[hotkey.lower()]}'",
         ]
-        if hotkey in ["C", "T"]:
+        if hotkey != hotkey.lower():
             event_parts.append("shiftKey: true")
         script = f"""jQuery.event.trigger({{
           {', '.join(event_parts)}
         }});"""
+        # print(script, flush=True)
         # pylint: disable=protected-access
-        el = self.browser.find_by_tag("body")
+        el = self.browser.find_by_id("thetext")
         self.browser.execute_script(script, el._element)
         time.sleep(0.2)  # Or it's too fast.
         # print(script)
-        # Have to refresh the content to query the dom ...
-        # Unfortunately, I can't see how to refresh without reloading
-        self.browser.reload()
-        time.sleep(0.2)  # Hack, test failing.
+        # Have to refresh the content to query the dom.
+        self._refresh_browser()
 
     def click_word_fill_form(self, word, updates=None):
         """
@@ -452,11 +533,9 @@ class LuteTestClient:  # pylint: disable=too-many-public-methods
             if "updated" in iframe.html:
                 should_refresh = True
 
-        # Have to refresh the content to query the dom ...
-        # Unfortunately, I can't see how to refresh without reloading
+        # Have to refresh the content to query the dom.
         if should_refresh:
-            self.browser.reload()
-            time.sleep(0.2)  # Hack, test failing.
+            self._refresh_browser()
 
     ################################3
     # Misc.